+++ src/Compiler/IL/LIRToILCompiler.MethodBodyCompilation.cs
@@ -237,20 +237,6 @@ internal sealed partial class LIRToILCompiler
                     var ctorDef = (MethodDefinitionHandle)token;
 
                     int argc = newUserClass.Arguments.Count;
-                    if (argc < newUserClass.MinArgCount || argc > newUserClass.MaxArgCount)
-                    {
-                        var expectedMinArgs = newUserClass.MinArgCount;
-                        var expectedMaxArgs = newUserClass.MaxArgCount;
-
-                        if (expectedMinArgs == expectedMaxArgs)
-                        {
-                            ILEmitHelpers.ThrowNotSupported(
-                                $"Constructor for class '{newUserClass.ClassName}' expects {expectedMinArgs} argument(s) but call site has {argc}.");
-                        }
-
-                        ILEmitHelpers.ThrowNotSupported(
-                            $"Constructor for class '{newUserClass.ClassName}' expects {expectedMinArgs}-{expectedMaxArgs} argument(s) but call site has {argc}.");
-                    }
 
                     if (newUserClass.NeedsScopes)
                     {