+++ src/Compiler/IR/HIR/HIRPrivateBrandCheckExpression.cs
@@ -1,0 +1,7 @@
+namespace Jroc.HIR;
+
+public sealed class HIRPrivateBrandCheckExpression : HIRExpression
+{
+    public required string RegistryClassName { get; init; }
+    public required HIRExpression Value { get; init; }
+}